refactor: create movement_display module, add relevant items to it
