+++ src/main.rs
@@ -10,6 +10,7 @@ mod gravestone;
 mod history;
 mod level_select;
 mod level_transition;
+mod movement_display;
 mod movement_table;
 mod nine_slice;
 mod previous_component;
@@ -109,7 +110,7 @@ fn main() {
         .add_system_set(
             ConditionSet::new()
                 .run_in_state(GameState::LevelTransition)
-                .with_system(gameplay::transitions::spawn_control_display)
+                .with_system(movement_display::spawn_control_display)
                 .with_system(gameplay::transitions::spawn_goal_ghosts)
                 .into(),
         )
@@ -132,14 +133,13 @@ fn main() {
         .add_system(gameplay::transitions::spawn_death_card.run_in_state(GameState::Gameplay))
         .add_system_to_stage(
             CoreStage::PreUpdate,
-            gameplay::systems::update_control_display.run_in_state(GameState::Gameplay),
+            movement_display::update_control_display.run_in_state(GameState::Gameplay),
         )
         .add_system(sugar::goal_ghost_animation.run_not_in_state(GameState::AssetLoading))
         .add_system(sugar::goal_ghost_event_sugar.run_not_in_state(GameState::AssetLoading))
         .add_system(sugar::animate_grass_system.run_not_in_state(GameState::AssetLoading))
         .register_ldtk_entity::<bundles::GoalBundle>("Goal")
         .register_ldtk_entity::<bundles::GrassBundle>("Grass")
-        .register_ldtk_int_cell::<bundles::ExorcismBlockBundle>(2)
         .register_ldtk_int_cell::<bundles::ExorcismBlockBundle>(2);
 
     #[cfg(feature = "hot")]
